requirements.txt
absl-py==2.1.0
accelerate==0.18.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip==0.2.0
cmake==3.28.1
confection==0.1.4
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
diffusers==0.25.0
docker-pycreds==0.4.0
einops==0.7.0
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl#sha256=8902305d2ced83d98a8e88efd93ea8970a70dda9bb24b0024a8b798d1cc913d5
executing==2.0.1
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.1
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.1
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.7.1
langcodes==3.3.0
lightning-utilities==0.10.1
lit==17.0.6
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
natsort==8.4.0
networkx==3.2.1
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pathy==0.10.3
pexpect==4.9.0
Pillow==10.1.0
plotly==5.18.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
pyDeprecate==0.3.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.5.2
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.2.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.1.2
torchmetrics==1.3.0.post0
torchvision==0.16.2
tqdm==4.65.0
traitlets==5.14.1
transformers @ git+https://github.com/huggingface/transformers.git@dbc12269ed5546b2da9236b9f1078b95b6a4d3d5
triton==2.1.0
typer==0.7.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
wandb==0.16.1
wasabi==1.1.2
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
ultralytics==8.2.12
supervision==0.20.0
scikit-image==0.23.2
inflect==7.2.1
